+++ supervisor/install.sh
@@ -1,15 +1,17 @@
 #!/bin/bash
 DIRNAME=$(dirname ${BASH_SOURCE[0]})
 
-yum install python-pip -y
-pip install supervisor
+if [ ! -e "/etc/supervisord/" ]; then
+	yum install python-pip -y
+	pip install supervisor
 
-mkdir -p /etc/supervisord/conf.d
-echo_supervisord_conf > /etc/supervisord.conf
-echo "[include]" >> /etc/supervisord.conf
-echo "files = /etc/supervisord/conf.d/*.conf" >> /etc/supervisord.conf
+	mkdir -p /etc/supervisord/conf.d
+	echo_supervisord_conf > /etc/supervisord.conf
+	echo "[include]" >> /etc/supervisord.conf
+	echo "files = /etc/supervisord/conf.d/*.conf" >> /etc/supervisord.conf
 
-cp ${DIRNAME}/supervisord.service /usr/lib/systemd/system/supervisord.service
+	cp ${DIRNAME}/supervisord.service /usr/lib/systemd/system/supervisord.service
 
-systemctl start supervisord
-systemctl enable supervisord
+	systemctl start supervisord
+	systemctl enable supervisord
+fi
